BIG Push, lots of features org logo will display on back of event for easier identification.
added timezone selection as well as ip and JS based.

fixed edit/delete menu to show for creator.

added extra event fields to edit method as well as create.

small changes, made edit / delete dropdown click instead of hover
added the mission brief url to view page if exists.

changed up the ip detection to a new more reliable API.

backend cleaning of code to stop duplication of code.

changed date number color so that it can be determined on a day with an event.

Signed-off-by: arthmael8610 <arthamel86@gmail.com>